motion_ai/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
opencv-python>=4.8.0
Pillow>=10.0.0
h5py>=3.9.0

# Visualization & Dashboard
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Metrics & Evaluation
scikit-learn>=1.3.0
pandas>=2.0.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Model Export
onnx>=1.15.0
onnxruntime>=1.16.0

# Utilities
tqdm>=4.66.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
ipywidgets>=8.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
